virt: expose source_file disk property to user
User may want to update or create a VM with a disk image created or
provided by a third party tool. In order to do this, let the user set
a source_file property to the disk.

Thus a {'name': 'myimage', 'source_file': '/path/to/image'} disk will
just use the provided disk image.