virt: changing disk bus should update it in the VM
So far two disks are considered equal if they have the same source file.
When a user only wants to change the disk bus, then the disk comparison
algorithm will consider it unchanged.

Now disks are compared on source and target bus. The other disk
variables can be left out side the target device is computed.