virt.update: refactor the XML diffing code
In order to handle more parameters in the future, refactor the XML
diffing in virt.update to only require adding entries in a mapping
array.

Using the refactoring the boot parameters are also better handled and we
are thus fixing an issue when updating existing loader parameter. See
the added test case.